+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Bugfixes:
 
   - Use #underscore, not #downcase for help text scope (#140, @atipugin)
   - Radio button and checkbox labels will now include the disabled class as needed. (#156, @ScottSwezey)
+  - Fixed issue with setting offset in form_group without label in horizontal layout  
+    when form uses non-default label_col
 
 Features: